@@ -1,7 +1,6 @@
 ﻿//------------------------------------------------------------------------------
 // <auto-generated>
 //     This code was generated by a tool.
-//     Runtime Version:4.0.30319.34014
 //
 //     Changes to this file may cause incorrect behavior and will be lost if
 //     the code is regenerated.
@@ -79,7 +78,7 @@ namespace EgoJpkArchiver.Properties {
         }
         
         /// <summary>
-        ///   Looks up a localized string similar to 2.0.
+        ///   Looks up a localized string similar to 2.1.
         /// </summary>
         internal static string AppVersionShort {
             get {
